backend/api/pyproject.toml
[tool.poetry]
name = "fapi"
version = "0.1.0"
description = ""
authors = ["Donovan Dicks <donovan.dicks@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.70.0"
spotipy = "^2.19.0"
python-dotenv = "^0.19.2"
uvicorn = "^0.15.0"
pytest = "^6.2.5"
pydantic = "^1.8.2"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
pylint = "^2.12.1"
black = "^21.11b1"

[tool.black]
line-length = 100
target-version = ["py310"]
fast = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
